Update enchantment-upgrade-tables.md
Removed the Feather Falling enchantment from the list as it cannot go above vanilla. Also added a note about how you are not able to use a crossbow on the upgrade enchant table.

I have been unable to test whether the soul speed and swift sneak enchantments are correct, or if you can use the mace in the upgrade enchant table.

Signed-off-by: Muffin949 <224354047+Muffin949@users.noreply.github.com>